Display stop numbers with decimals when they exist, and as whole numbers otherwise
Modify `services_reconciliation.py` to store `stop_seq` as float and update templates (`shipment_detail.html`, `route_reconciliation.html`, `route_reconciliation_print.html`) to format `stop_seq` for display, showing decimals for values like 9.5 and whole numbers for values like 9.

+++ services_reconciliation.py
@@ -119,7 +119,7 @@ def get_invoice_reconciliation_report(shipment_id: int) -> List[Dict]:
         stop_id = r['route_stop_id']
         if stop_id not in stops_data:
             stops_data[stop_id] = {
-                'stop_seq': int(r['stop_seq'] or 0),
+                'stop_seq': float(r['stop_seq'] or 0),
                 'customer_name': r['customer_name'] or '',
                 'invoices': [],
                 'total_received': 0,
